Use internal package for shared code
Make util.ExchangeTKEY not require the optional TSIG MAC but require the
passed dns.Client to be configured beforehand.
